Added comments to sys.c custom UP() and DOWN() methods
Comments added to give a more high-level view of what exactly is
going on. The DOWN and UP methods are implemented as syscalls
because they use locks that disable hardware interrupts. This can
only be done in the kernel. When hardware interrupts are disabled,
the execution of code within these methods will run distinctly from
all other processes. This is what we call a CRITICAL REGION.