Update prodcons.c
Fixed shared memory space to be the exact amount of space needed.
Before this fix, the mapping was too large because all of the
resources were calculated and then multiplied together instead of
being added together.

This is a maintenance fix that should not affect the outcome of
the application. See line74 for exact mappings of this space.